Refactor OTel health contracts into atm-core
